Move all ParseTest tests and helpers to Unittest
* Move the files.
* Move all the classes into the `Psecio\Parse\Unittest` namespace.
* Tell Composer to add the tests directory as a search path for the `Psecio`
  namespace for dev using `autoload-dev`.
* Remove the `require` statement from `ParseTest.php` since autoloading now
  works.
