+++ composer.json
@@ -27,6 +27,11 @@
             "Psecio": "src/"
         }
     },
+    "autoload-dev": {
+        "psr-0": {
+            "Psecio": "tests/"
+        }
+    },
     "config": {
         "bin-dir": "bin"
     },
